pyproject.toml
[project]
name = "audio-web-app"
version = "0.1.0"
description = "Audio recording web app for collecting voice samples"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
]
